Allow tests to be run inside a docker container
This adds the `ip_address` method to `Docker` trait in order to
to query the ip address of the spawned container. This is derived
from a `docker inspect` call.

This also introduces a new usage pattern when used from the host system:
Rather than relying on a port mapping to the host, the target port
can be directly queried by using the ip address of the container in the
internal docker network.
Note: This relies on the fact, that the default docker bridge network is
configured for all involved containers.

https://github.com/testcontainers/testcontainers-rs/issues/129