Upgrade default bitcoin image version to 0.21.0
With the upgrade to 0.21.0 a new startup command was introduced which allows us to execute a script when the startup sequence was completed. We simply echo `bitcoind startup sequence completed.`.
This allows us to remove `-debug` which keeps the bitcoind log _clean_.
Note: this release also removed the default wallet.