Make `Docker` trait pub(crate)
To make this possible, we define the two `run` functions on the client
directly instead of enforcing it through an interface.

This allows us to make `Container` depend on a trait object instead
of using a type parameter. This simplifies the API and at the same
time allows us to shrink the public API of the crate, which reduces
breaking changes in the future.