Use `bollard-stubs` directly for models
`bollard-stubs` is a crate with autogenerated models for the Docker API.
Previously, we were using models embedded in `bollard` inside CLI, while
the client itself wasn't required. Now we're making `bollard` an
optional dependency and including it only when we need the
heavy-lifting, i.e. the client logic.

The internal implementation of the CLI client uses the models from
`bollard-stubs` from now on.