Add support to expose ports in GenericImage
By default, testcontainers-rs runs docker with `-P`. However, for a user
to get the port mappings, this relies on the image itself properly
exposing ports using the EXPOSE instruction [1]. When images do not use
this instruction, a user can manually expose these ports using the
`--expose` option in the docker CLI.

However, testcontainers-rs does not allow users to specify this argument
in GenericImage. This commit exposes a new method, with_exposed_port()
as part of a GenericImage builder that allows a user to properly expose
ports for any generic image.

[1]: https://docs.docker.com/engine/reference/builder/#expose
Closes: https://github.com/testcontainers/testcontainers-rs/issues/348