Add testimages crate for co-locating test images
This adds a testimages crate which can contain the source code for test
docker images.

Maintaining images within the repository gives nice advantages.
Depending on external repositories makes it unclear where the image is
coming from and who is meant to maintain it. It also bears the risk of
our CI breaking if the image ever disappears.

This patch sets up the structure that allows us to use the images
without having to push them anywhere and we can refer to them by name in
our tests.

New images can be added as new binaries in testimages/src/bin/, with
their corresponding Dockerfiles in testimages/src/dockerfiles/. Then,
build.rs would be updated to build that particular image.

Ref: https://github.com/testcontainers/testcontainers-rs/pull/349