Handle IPv4 and IPv6 host ports separately
This introduces new methods for getting host ports mapped from
containers, specific to IPv4 or IPv6 addresses. The existing methods are
deprecated in favor of their replacements.

When Docker publishes all of a container's exposed ports, it does so on
both IPv4 and IPv6. Port numbers are chosen separately for each, from
the ephemeral port range, though it is common for Docker to assign the
same host port number for both the IPv4 and IPv6 bindings of one exposed
port. However, if Docker encounters a port that is in use, it may choose
another.

Previously, get_host_port() would return the first port binding it found
in Docker's list for each container port. The new methods filter the
list by IP protocol version. As Docker tended to return port bindings
for 0.0.0.0 first, the now-deprecated get_host_port() forwards to
get_host_port_ipv4().