Actually map exposed ports for HTTP client (#531)
When switching from the Cli client to the Http client I noticed that the
expose_ports of my Image were no longer dynamically mapping to free
ports on the docker daemon's host. It seems like this is a gap to fill
as a user might want a specific container port (that the image manifest
may or may not have an EXPOSE layer for) mapped to a free port on the
daemon host.