feat(cli): support for docker cli flag `--add-host`
When spawning multiple docker containers in the same test,
communication amongst them can achieved over the docker default
gateway.

In DinD environments such as when running jobs on a CI pipeline, the docker
[`--add-host`](https://docs.docker.com/engine/reference/commandline/container_run/#option)
can be employed to basically enter new entries in the `/etc/hosts`
files. This allows to avoid binding all container ports to the host
environment.

The flag API counterpart is implemented on the `RunnableImage` struct
instead of on the generic image since I reckon it makes more sense
within this crate phylosophy, I hope !-).

A test runs 2 containers:

- an http_server
- a `curl` client

and the `curl` client completes when receiving `foo` from the server on the
alias hostname. In case the hostname wasn't resolve it'd panic.