Include LICENSE file in published crates
Apache-2.0 license term include a requirement that the (re)distributed
sources contain a copy of the license text. Similar conditions apply to
the MIT one. This is a problem for Linux distributions wanting to
include and package the crates from this projecs, affecting also
creates.io

This PR includes a symlink to the LICENSEs and README files in the
crate, which should make `cargo publish` include these files when
publishing to crates.io.

Signed-off-by: Daniel Mellado <dmellado@redhat.com>