feat: add `get_hostname` to containers and deprecate `get_host_ip_address`
Existed logic was incorrect, because we may encounter domain name.
Also, `IpAddr` isn't suitable for URL usage (because of `IpV6`).

url::Host is better alternative and supports:
- Domain names
- escapes `IpV6` for `URL`

We preserve `get_host_ip_address` for now, but we have to get rid of this method and extra dependency in the next major release