refactor!: rename `RunnableImage` to `ContainerRequest`
I decided to go with renaming because it's a good moment: we've provided `ImageExt` trait in order to avoid direct usage of `RunnableImage` when possible. Renaming seems to be minor issue in contrast of other changes.

This is semantically more correct, it allows to customize some parameters of a container to create, but not the image itself (`network`, `shm_size`, `cgroupns` etc.).
Subjectively, it's more obvious to see  `ContainerRequest<Redis>` instead of `RunnableImage<Redis>`.
Also, it's similar to Testcontainers for Go.