perf: use separate dropper thread for `current_thread` flavor (#666)
`tokio::test` run with `current_thread` flavor by default and it's
typical use-case.

However, we need to perform some async operations on drop, and it used
to spawn a new thread for each drop. It's inefficient and was a
temporary solution.

Now, it uses:
- for `multi-thread` runtime flavor: `tokio::task::block_in_place`
- for `current-thread` runtime flavor: lazily spawns a separate thread
for dropping tasks, thus it will be only 1 thread regardless of the
number of containers, networks, etc.