fix!: drop `Clone` impl for `ContainerRequest` (#680)
`Image` can be cloned just fine, but `ContainerRequest` is not intended
to be, and is not, a truly safe structure for cloning. It allows to
override unique container properties: container name, port mappings and
log consumers (soon).

If necessary, you can always use a reusable constructor/method to create
multiple similar `ContainerRequests`.