feat: introduce log consumers (#681)
Introduce `LogConsumer` trait and ability to pass a consumer into
`ContainerRequest`.

The main advantage of this functionality is the use of background
container output processing throughout the entire lifecycle of the
container, starting from the moment of its creation. This way you can
process logs even if the container was not started due to the wait
strategy.