feat: add `working_dir` to `ContainerRequest`,`ImageExt` (#724)
Implements #720 by adding support for setting the working directory on
`ImageExt` and `ContainerRequest`. Also adds a test for this feature.
This is equivalent to the `-w, --workdir` flag of `docker run`.

The abbreviation `working_dir` (as opposed to e.g. `working_directory`)
was chosen to use the same name as the `bollard` crate.