feat: expose configured bollard client
This PR allows to access bollard client directly if needed.
It's aligned with implementations of testcontainers for other languages. Definitely may be useful in some specific cases and allows to bypass some limitations.
We could wrap and expose only particular API methods and hide the details from the user, but it's too much work and this way seems more pragmatic.

Later we also can re-export bollard and bollard_stubs on demand (in order to simplify version managment of deps for users)