Add some more documentation and a doctest on image creation
I struggled a bit when upgrading to testcontainers-rs 0.23, since the
interfaces had changed somewhat since the last version I used. The
documentation didn't explain exactly how to use the methods in `ImageExt`
in conjunction with those of `GenericImage`.

This change adds a bit more content to the documentation:

* A few more lines in the README file to exercise the configuration
  methods in `ImageExt`,
* Some documentation for `GenericImage` showing how it is used along with
  `ImageExt`.

The latter in particular includes a doctest. If the API is changed in the
future so that further changes are needed, the doctest will fail and can
be updated accordingly. This will make it easier for downstream users to
understand how to port their code.