feat: add option to stop containers with a timeout (#779)
Previously containers were stopped with no timeout (SIGKILL) which was
interfering with clean shutdown of processes being tested.

This commit adds a new API of stop_with_timeout(timeout_secconds) in
addition to the original stop(). The original stop() interface calls
stop_with_timeout using the default '0'.

EDIT:

It would seem my original diagnosis was incorrect and the default
timeout is 10 seconds, but my CI system was somehow setting the default
to 1 second. I will fix the docs and put in the correct default. I think
there is still some value in here - apologies for the time waste getting
to this point!