chore(deps): drop weak dependencies due to cargo bug (#862)
For now we're going to drop weak deps for russh (`russh?/ring` and
`russh?/aws-lc-rs`) as there is a known bug with Cargo which pulls it to
the Cargo.lock and may trigger cargo-audit for example (e.g known `rsa`
vulnerability). See details in [this
comment](https://github.com/testcontainers/testcontainers-rs/issues/851#issuecomment-3452202143)

With this change, we simply require `russh/ring` for
`host-port-exposure` feature.

As a workaround, users who needs host-exposure feature with aws-lc-rs
can add russh to their deps and add it explicitly in their project.
However it may require to install the CryptoProvider.