feat: support for reusable containers that have been stopped (#896)
This PR allows the `reusable-containers` feature to reuse stopped
containers.

## Motivation

I am trying to write integration tests with ClickHouse and Postgres, but
I don't want to spin up one container per test. Moving the containers to
a static OnceCell works, but unfortunately static variables are not
dropped so the containers are not destroyed.

Integrating Ryuk as a solution to works, but the Ryuk container is not
able to remove itself. Ryuk will successfully remove all test
containers, and then shutdown. In order to avoid generating a Ryuk
container per test run, I'd like to use reusable containers.
Unfortunately, the current code fail as soon as Ryuk shuts down.

I am actually happy that Ryuk shuts down since that means it won't hog
resources. What I want is for the reusable containers to be able to
reuse my stopped Ryuk container, since that is the cleanest solution I
have found for test container management on OnceCells.